+++ src/main/java/com/paqueteria/model/TarifaDistancia.java
@@ -1,5 +1,6 @@
 package com.paqueteria.model;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -30,7 +31,7 @@ public class TarifaDistancia {
     
     @NotNull(message = "El coste no puede ser nulo")
     @Column(nullable = false)
-    private Float coste;
+    private BigDecimal coste;
     
     @NotNull(message = "El estado activa no puede ser nulo")
     @Column(nullable = false)
@@ -42,7 +43,7 @@ public class TarifaDistancia {
     // Constructors
     public TarifaDistancia() {}
     
-    public TarifaDistancia(DistanciaEnum distancia, Float coste) {
+    public TarifaDistancia(DistanciaEnum distancia, BigDecimal coste) {
         this.distancia = distancia;
         this.coste = coste;
         this.activa = true;
@@ -61,11 +62,11 @@ public class TarifaDistancia {
         this.distancia = distancia;
     }
     
-    public Float getCoste() {
+    public BigDecimal getCoste() {
         return coste;
     }
     
-    public void setCoste(Float coste) {
+    public void setCoste(BigDecimal coste) {
         this.coste = coste;
     }
